+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Clobbering behaviour
 - Possibility to create destination directory on the fly
 - Possibility to specify directory using linux path syntax
+- Possibility to specify base directory
 
 ### Changed
 - Small refactoring